--- components/experimental/tests/transliterate/data/baked/any.rs
@@ -1,2 +1,0 @@
-// @generated
-impl_any_provider!(BakedDataProvider);
