+++ components/experimental/tests/transliterate/data/baked/fingerprints.csv
@@ -1,0 +1,1 @@
+transliterator/rules@1, <lookup>, 320B, 16 identifiers
